requirements/base.txt
Django==5.0.2
django-cors-headers==4.9.0
django-crontab==0.7.1
django-ninja==1.5.1
django-ninja-extra==0.30.8
django-redis==6.0.0
django-redisboard==9.0.0
django-salesforce-agpl==5.2
django-openstax-accounts==1.0
django-openstax-healthcheck==1.0
psycopg2==2.9.11
PyJWE==1.0.0
PyJWT==2.10.1
python-dotenv==1.2.1
sentry-sdk==2.49.0
simple_salesforce==1.12.5
